feat: 관리자 API 분리 및 인증/세션/JWT 로그아웃(블랙리스트) 기능 통합
* 관리자(Admin) API 구조를 도메인 기준으로 분리
  * 유저/디바이스/세션/로그인시도 컨트롤러 및 서비스(Query/Command) 책임 분리
  * 라우팅 및 의존성 명칭 정리로 유지보수성 개선

* 관리자 유저 관리 기능 확장
  * 유저 목록 검색/필터 지원(q/email/nickname/provider/suspended/가입일 범위)
  * 계정 정지/해제(사유/기간 저장) 및 유저 프로필 부분 수정(PATCH) 추가
  * UserAccount 정지 관련 필드(suspendedAt/suspendedUntil/suspendedReason) 및 스펙(UserAccountSpecs) 추가

* 관리자 디바이스/세션 관리 기능 확장
  * 유저 디바이스 목록 조회 API 추가
  * 디바이스 단위 활성 세션 강제 종료(revoke) API 추가
  * 유저 세션 목록 조회(activeOnly/deviceId) 및 세션 단건 상세 조회 API 추가

* 로그인 시도(로그) 관리 기능 확장
  * 로그인 시도 단건 상세 조회 API 추가
  * 기간/필터 기반 통계(stats) API 추가(성공/실패/Provider/에러코드/Top IP·Device)
  * Repository 통계 쿼리 및 DTO 확장

* JWT 로그아웃(블랙리스트/denylist) 도입 및 만료 정리 배치 추가
  * jwt_denylist 엔티티/리포지토리/서비스(revoke/isRevoked/cleanupExpired) 추가
  * RevocationAwareJwtDecoder로 revoked 토큰 차단
  * JWT 로그아웃 API(/api/v1/auth/jwt/logout) 추가
  * 만료된 denylist 정리 스케줄러(@Scheduled) 추가 및 스케줄 활성화

* 세션 로그아웃 처리 보강
  * 로그아웃 시 user_session_map revoke 기록(사유/시간) + Spring Session 삭제로 즉시 무효화
  * 세션 기반 로그아웃과 JWT 기반 로그아웃 역할 분리 명확화
* 보안 설정 보완
  * /api/** 체인에서 인증 예외 경로(회원가입/로그인 등) permitAll, JWT 로그아웃은 authenticated 유지하도록 정리

+++ core/core-api/src/main/kotlin/io/soo/springboot/core/api/controller/v1/response/AdminUserSummaryDto.kt
@@ -12,5 +12,5 @@ data class AdminUserSummaryDto(
     val updatedAt: LocalDateTime,
     val providers: List<AuthProvider>,
     val deviceCount: Int,
-    val activeSessionCount: Long,
+    val activeSessionCount: Int,
 )
